memcached/app/search/requirements.txt
python-memcached===1.59
uhashring==2.3
numpy==1.21.5

